Solve PR comments
