+++ RealmSwift/Realm.swift
@@ -1512,3 +1512,43 @@ extension Projection: RealmFetchable {
         return Root.className()
     }
 }
+
+/**
+ `Logger` is used for creating your own custom logging logic.
+
+ You can define your own logger creating an instance of `Logger` and define the log function which will be
+ invoked whenever there is a log message.
+
+ ```swift
+ let logger = Logger(level: .all) { level, message in
+    print("Realm Log - \(level): \(message)")
+ }
+ ```
+
+ Set this custom logger as you default logger using `Logger.shared`.
+
+ ```swift
+    Logger.shared = inMemoryLogger
+ ```
+
+ - note: By default default log threshold level is `.info`, and logging strings are output to Apple System Logger.
+*/
+public typealias Logger = RLMLogger
+extension Logger {
+    /**
+     Log a message to the supplied level.
+
+     ```swift
+     let logger = Logger(level: .info, logFunction: { level, message in
+         print("Realm Log - \(level): \(message)")
+     })
+     logger.log(level: .info, message: "Info DB: Database opened succesfully")
+     ```
+
+     - parameter level: The log level for the message.
+     - parameter message: The message to log.
+     */
+    internal func log(level: LogLevel, message: String) {
+        self.logLevel(level, message: message)
+    }
+}
