+++ RealmSwift/Tests/RealmTests.swift
@@ -1946,7 +1946,7 @@ class LoggerTests: TestCase {
         Logger.shared.level = .all
         try autoreleasepool { _ = try Realm() } // We should be getting logs after changing the log level
         XCTAssertEqual(Logger.shared.level, .all)
-        XCTAssertTrue(logs.contains("Info DB:"))
+        XCTAssertTrue(logs.contains("Details DB:"))
         XCTAssertTrue(logs.contains("Trace DB:"))
     }
 
@@ -1962,13 +1962,13 @@ class LoggerTests: TestCase {
         XCTAssertTrue(logs.isEmpty)
 
         // Info
-        Logger.shared.level = .info
+        Logger.shared.level = .detail
         try autoreleasepool { _ = try Realm() }
 
         XCTAssertTrue(!logs.isEmpty)
-        XCTAssertTrue(logs.contains("Info DB:"))
+        XCTAssertTrue(logs.contains("Details DB:"))
 
-        // Debug
+        // Trace
         logs = ""
         Logger.shared.level = .trace
         try autoreleasepool { _ = try Realm() }
@@ -1976,13 +1976,13 @@ class LoggerTests: TestCase {
         XCTAssertTrue(!logs.isEmpty)
         XCTAssertTrue(logs.contains("Trace DB:"))
 
-        // Info
+        // Detail
         logs = ""
-        Logger.shared.level = .info
+        Logger.shared.level = .detail
         try autoreleasepool { _ = try Realm() }
 
         XCTAssertTrue(!logs.isEmpty)
-        XCTAssertTrue(logs.contains("Info DB:"))
+        XCTAssertTrue(logs.contains("Details DB:"))
         XCTAssertFalse(logs.contains("Trace DB:"))
 
         logs = ""
@@ -1992,31 +1992,7 @@ class LoggerTests: TestCase {
         XCTAssertEqual(Logger.shared.level, .trace)
         try autoreleasepool { _ = try Realm() }
         XCTAssertTrue(!logs.isEmpty)
-        XCTAssertTrue(logs.contains("Info DB:"))
+        XCTAssertTrue(logs.contains("Details DB:"))
         XCTAssertTrue(logs.contains("Trace DB:"))
-    }
-
-    func testCustomLoggerLog() throws {
-        var logs: String = ""
-        let logger = Logger(level: .info) { level, message in
-            logs += "\(Date.now) \(level.logLevel) \(message)"
-        }
-        Logger.shared = logger
-        XCTAssertTrue(logs.isEmpty)
-
-        try autoreleasepool { _ = try Realm() }
-
-        XCTAssertTrue(!logs.isEmpty)
-        XCTAssertNotNil(logs.contains("Info DB:"))
-
-        logger.log(level: .info, message: "Info DB: 'Database is broken'")
-        XCTAssertTrue(logs.contains("Info DB:"))
-        XCTAssertTrue(logs.contains("Info DB: 'Database is broken'"))
-
-        logger.log(level: .debug, message: "Debug DB: 'Database is good'")
-        // Should not be added to log, because log level is greater than the one set.
-        XCTAssertTrue(logs.contains("Info DB:"))
-        XCTAssertTrue(logs.contains("Info DB: 'Database is broken'"))
-        XCTAssertFalse(logs.contains("Debug DB: 'Database is good'"))
     }
 }
